Fix dtype conversion bug in probe model data loader
The data loader was yielding uint16 tensors directly from .bin files,
but PyTorch's embedding layer and cross_entropy loss require Long (int64) type.

Error encountered:
  RuntimeError: Expected tensor for argument #1 'indices' to have one
  of the following scalar types: Long, Int; but got CUDAUInt16Type instead

Fix:
- Convert input_ids to .long() before yielding (uint16 -> int64)
- Convert targets to .long() before yielding (uint16 -> int64)
- Embedding layer now receives correct dtype
- CrossEntropyLoss now receives correct dtype

This maintains memory efficiency (data stored as uint16) while ensuring
correct types for PyTorch operations.

+++ train_probe_model.py
@@ -192,8 +192,8 @@ def simple_data_loader(filename_pattern: str, num_tokens: int, seq_len: int, bat
         # Reshape to [batch_size, seq_len + 1]
         batch_tokens = batch_tokens.view(batch_size, seq_len + 1)
 
-        input_ids = batch_tokens[:, :-1]
-        targets = batch_tokens[:, 1:]
+        input_ids = batch_tokens[:, :-1].long()  # Convert uint16 -> int64 for embedding
+        targets = batch_tokens[:, 1:].long()    # Convert uint16 -> int64 for cross_entropy
 
         yield input_ids, targets
